requirements.txt
wandb 
pytorch-lightning
clip-anytorch
pySmartDL
transformers

pycocotools
torchattacks
h5py
matplotlib
numpy
pandas
Pillow
scipy
scikit-learn
tqdm
torchvision
gdown